[UPDATE] Apply technique for attackers

+++ Assets/Scripts/Pool/Arrows/Arrow.cs
@@ -13,9 +13,9 @@ public class Arrow : BasePoolMember
     private float _damageRadius = 1f;
 
     private Coroutine _fireCoroutine;
-    private int _damage = 1;
+    private float _damage = 1;
 
-    public void SetDamage(int damage) => _damage = damage;
+    public void SetDamage(float damage) => _damage = damage;
 
     private void OnDisable()
     {
